# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a group of API for WTM Service",
        "title": "WTM Backend Service",
        "contact": {},
        "version": "1"
    },
    "paths": {
...
        "/hotels": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a paginated list of hotels using query parameters.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hotel"
                ],
                "summary": "List Hotels",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Filter hotels by API status",
                        "name": "is_api",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "Filter hotels by region",
                        "name": "region",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter hotels by status_id",
                        "name": "status_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number for pagination (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search keyword to filter hotels by name",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved list of hotels",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseWithPagination"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/hoteldto.ListHotel"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hotel"
                ],
                "summary": "Create Hotel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hotel name",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "file"
                        },
                        "collectionFormat": "multi",
                        "description": "Hotel photos (multiple allowed)",
                        "name": "photos",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Sub-district location",
                        "name": "sub_district",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "District location",
                        "name": "district",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Province location",
                        "name": "province",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Contact email",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Hotel description",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Hotel rating (1–5)",
                        "name": "rating",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Nearby places as JSON string. Example: /example_nearby_places ",
                        "name": "nearby_places",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "Facilities (multiple allowed)",
                        "name": "facilities",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Social media links as JSON string. Example: /example_social_medias ",
                        "name": "social_medias",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully created hotel",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseWithData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/hoteldto.CreateHotelResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/hotels/agent": {
            "get": {
                "description": "Retrieve a list of hotels for a specific agent with pagination and filtering options.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hotel"
                ],
                "summary": "List Hotels for Agent",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number of hotels to return per page (default: 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number to retrieve (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term to filter hotels by name or description",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter hotels by province",
                        "name": "province",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "Filter hotels by district",
                        "name": "district",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "multi",
                        "description": "Filter hotels by rating (e.g., 0,1,2,3,4,5)",
                        "name": "rating",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Minimum price range for filtering hotels",
                        "name": "range_price_min",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Maximum price range for filtering hotels",
                        "name": "range_price_max",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "multi",
                        "description": "Filter hotels by total number of bedrooms (e.g., 1,2,3,4,5)",
                        "name": "total_bedrooms",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter hotels by total number of rooms (e.g., 1,2,3,4,5)",
                        "name": "total_rooms",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter hotels by total number of guests (e.g., 1,2,3,4,5)",
                        "name": "total_quests",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter hotels by promo id",
                        "name": "promo_id",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "multi",
                        "description": "Filter hotels by bed type Id (e.g., 1,2,3)",
                        "name": "bed_type_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved list of hotels for agent",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseWithPagination"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/hoteldto.ListHotelForAgentResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/promos/agent": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List promos for agent",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Promo"
                ],
                "summary": "List promos for agent",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number for pagination",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search keyword to filter promos by name,description, and code",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved list of promos",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseWithPagination"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/promodto.PromosForAgent"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.FilterBedTypeHotel": {
            "type": "object",
            "properties": {
                "bed_type": {
                    "type": "string"
                },
                "bed_type_id": {
                    "type": "integer"
                },
                "count": {
                    "type": "integer"
                }
            }
        },
        "entity.FilterRangePrice": {
            "type": "object",
            "properties": {
                "max_price": {
                    "type": "number"
                },
                "min_price": {
                    "type": "number"
                }
            }
        },
        "entity.FilterRatingHotel": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "rating": {
                    "type": "integer"
                }
            }
        },
        "entity.FilterTotalBedroom": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "total_bed_rooms": {
                    "type": "integer"
                }
            }
        },
...
        "hoteldto.CreateHotelResponse": {
            "type": "object",
            "properties": {
                "hotel_id": {
                    "type": "integer"
                }
            }
        },
...
        "hoteldto.ListHotel": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_api": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "rooms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/hoteldto.RoomTypeItem"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "hoteldto.ListHotelForAgent": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "min_price": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "photo": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer"
                }
            }
        },
        "hoteldto.ListHotelForAgentResponse": {
            "type": "object",
            "properties": {
                "filter_bed_types": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.FilterBedTypeHotel"
                    }
                },
                "filter_districts": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "filter_pricing": {
                    "$ref": "#/definitions/entity.FilterRangePrice"
                },
                "filter_ratings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.FilterRatingHotel"
                    }
                },
                "filter_total_rooms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.FilterTotalBedroom"
                    }
                },
                "hotels": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/hoteldto.ListHotelForAgent"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "hoteldto.RoomTypeItem": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "price_with_breakfast": {
                    "type": "number"
                }
            }
        },
...
        "promodto.PromosForAgent": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "hotel": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "response.Pagination": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "response.ResponseWithData": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "response.ResponseWithPagination": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/response.Pagination"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}